requirements_simple.txt
streamlit>=1.37.0
ultralytics>=8.3.0
opencv-python-headless>=4.10.0
numpy>=2.1.0
pillow>=10.4.0